+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 #### Builtin
 
+- **Breaking Change**: change ``OSTYPE`` definition, now ``OSTYPE`` is not equivalent to uname
+    - ``linux``, ``darwin``, ``cygwin``, ``emscripten``
 - show row numbers in completion pager if actual rows are larger than rendered rows
 
 ## [0.29.2] - 2023-04-08